Added support for objects which override the ToString method in SessionModelConverter
SessionModelConverter only had support for types marked as Serializable.
With this change, it also checks if the type has overridden the ToString
method in which case it calls it. If the type is neither marked as
serializable nor has the ToString method, then the "Non serializable
type" error message is shown in the Session tab.