Made the Request flow handling logging opt-in
Logging was added to the GlimpseRuntimeWrapper to log the events it was
processing and whether or not duplicates were found during processing.
It also logged calls made for resources etc... This needed to be
conditional, because the history tab is polling every second for its
resource, filling the glimpse.log at such a rapid pace that usefull
information got obscured. So an optional attribute has been added to the
glimpse logging element, allowing to opt-in for that kind of logging in
case someone needs it.