Test run of storing types
Working on Feedback

Attempt to make it more performant from comments

Limit access to dictionary and add locks

Check if in GAC

This pretty much eliminates this as a performance concern.  It gets rid
of 50 of the dlls, and the biggest ones.

Revert "Check if in GAC"

This reverts commit fd94dff1ac6e75a8bc218f843442a6fae9aa6296.

Add a little meat to the class

Add black list

I may have gone a little overboard on blacklisting things.

refactor dictionary name

Move Lock to outer scope

Move classes to their own files

Move files to correct location

Move lock to outer scope.

New line

And clean up Glimpse.Core.Net45 to be able to compile