Fixes #631 Made Glimpse.axd CSP Compliant
- Added a test page for the Glimpse Client and that is still not CSP
Compliant especially due to styles being inlined.
- To test this I've added the "Content-Security-Policy", = "default-src
'self'" header for the glimpse.axd and that test page in the MVC3 Music
Store Sample. Although it is to restrictive for the test page containing
the Glimpse panel. To make that work "style-src 'unsafe-inline'" must be
added to the header as well.