Made sure Glimpse.Ado works again with GlimpseRuntime.Instance
Glimpse.Ado is activated by means of the PreApplicationStartMethod, so
that is injected as soon as possible. Unfortunately the Glimpse
HttpModule did not yet initialize the GlimpseRuntime, which will throw a
GlimpseNotInitializedException when GlimpseRuntime.Instance is being
requested by Ado code. This fix will handle that exception and fall back
to the v1 situation, meaning null will be returned for the
IMessageBroker and the IExecutionTimer, and those null values were
already dealt with.