Fix #646 - Ugly hack to get around bug in EF6
EF6 has a bug which is described here -
http://stackoverflow.com/questions/19966106/spatial-datareader-and-wrapping-providers-in-ef6.
This was picked up and fixed in EF 6.1 here -
https://entityframework.codeplex.com/workitem/1867. Because we don't
release a new version of Glimpse.EF for point releases of EF, we have to
go with this hack until we hit EF7. So we don't forget the code that is
needed is already in place.