Ajax requests that are using "Bufferless input streams" cannot be monitored
HttpRequest.Files throws an excpetion if the request is read using a
bufferless input stream (this is the case for some of my WCF services
with JSON endpoint). The excpetion screws up the whole switching to that
request. Catching the error will show all other data (e.g. SQL
statements) for that request without problem.